Renaming the files
